requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
reportlab==4.0.9
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.9.3
numpy==1.26.0
pandas==2.1.1
